責務が薄い2クラスを統合。FileListPreparer + MyFile = FileList
